Refactor/#88 : Gateway 리팩토링
