+++ server/gateway/src/api/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import { Global, Module } from '@nestjs/common';
 import { ClientsModule, Transport } from '@nestjs/microservices';
 
-import { MICRO_SERVICES } from '../constants/microservices';
+import { MICRO_SERVICES } from '../../common/constants/microservices';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
