+++ server/gateway/src/api/ticket/ticket.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ClientsModule, Transport } from '@nestjs/microservices';
 
-import { MICRO_SERVICES } from '../constants/microservices';
+import { MICRO_SERVICES } from '../../common/constants/microservices';
 import { TicketController } from './ticket.controller';
 
 @Module({
